Fix broken error-handling test and add success-path integration test
testInvalidFileTypeReturnsJsonError was failing at the CSRF guard
(returning 400) before ever reaching the extension check it claimed to
test — it only passed because the error envelope shape happened to
match. It now mints a valid CSRF token and asserts 422 from the
extension check.

Adds testMissingCsrfTokenReturnsJsonError to preserve coverage of the
400-without-token path, and testValidUploadReturnsCsvDownload to close
the gap of having no success-path integration test anywhere in the
suite.

Whitelists 127.0.0.1 in .env.test so integration tests don't trip the
5/10-min rate limit across runs.

Renames the test class to TransformControllerIntegrationTest since it
now covers both error and success paths.
